requirements.txt
aiofiles==0.5.0
certifi==2020.4.5.2
chardet==3.0.4
dnspython==1.16.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.6.9
httptools==0.1.1
httpx==0.11.1
hyperframe==5.2.0
idna==2.9
marshmallow==2.21.0
motor==2.1.0
multidict==4.7.6
pymongo==3.10.1
python-dateutil==2.8.1
rfc3986==1.4.0
sanic==20.3.0
sanic-base-extension==0.1.1
sanic-mongodb-extension==0.4.0
six==1.15.0
sniffio==1.1.0
umongo==2.2.0
urllib3==1.25.9
websockets==8.1
requests==2.24.0
Sanic-Cors==0.10.0.post3
swagger-ui-py==0.3.0
PyJWT==1.7.1
pytz==2020.1
python-docx==0.8.10
pandas==1.2.3
nltk==3.5.0
seaborn==0.11.1
# tornado